feat: zoom to layer and close menu when on mobile, refactor the useSidebar to lift state up into a sidbar context (#54)
* feat: zoom to layer and close menu when on mobile, refactor the useSidebar to lift state up into a sidbar context

* feat: enable the zoom to button to also trigger the group layer to be clicked
