+++ src/components/app-shell.tsx
@@ -1,12 +1,12 @@
 import { Outlet } from 'react-router-dom'
 import Sidebar from './sidebar'
-import useIsCollapsed from '@/hooks/use-is-collapsed'
+import { useSidebar } from '@/hooks/use-sidebar';
 
 export default function AppShell() {
-  const [isCollapsed, setIsCollapsed] = useIsCollapsed()
+  const { isCollapsed } = useSidebar();
   return (
     <div className='relative h-full overflow-hidden bg-background'>
-      <Sidebar isCollapsed={isCollapsed} setIsCollapsed={setIsCollapsed} />
+      <Sidebar />
       <main
         id='content'
         className={`overflow-x-hidden pt-16 transition-[margin] md:overflow-y-hidden md:pt-0 ${isCollapsed ? 'md:ml-14' : 'md:ml-[36rem]'} h-full`}
