+++ src/hooks/use-check-active-nav.tsx
@@ -3,7 +3,6 @@ import { useSidebar } from '@/hooks/use-sidebar'
 export default function useCheckActiveNav() {
   const { currentContent } = useSidebar()
 
-
   const checkActiveNav = (nav: string) => {
     if (!currentContent) return false
